Deeper recursion for commit 471b03b39
Refer to that commit's message.
The extension made is the following:
If a top is not available, navigate up its pre chain until there is a
node that is available.

One more extension is evident:
Instead of just navigating up the node's column,
also navigate up the node's top's pre's columns.

That's a lot more recursion and will be left for a later date.