Major overhaul cont from commit e187a86624
Previously, _valid_tops would check if modifying a node's
pre-connections would change that node's pg status. If it did, that
would be undesirable and rejected.

But actually, what one cares about is not that specific node, but rather
its effects on the post-processing nodes. If the post-processing nodes
already get their pg from somewhere else, the node's pg status may
change and still leave the transform valid!

So a scheme must be devised to check the post-processing nodes, and do
it efficiently.
_remains_pg_valid does this beautifully.