Add "shift" methods
These right_shift and left_shift methods will attempt to
right_rotate or left_rotate a node even when such a rotation is not
possible.

This is performed by rotating the node the wrong way until the
correct-way rotation can be performed.

This is strictly analogous to LR or RL combined rotations in AVL tree
balancing, except generalized, since AVL trees don't have to deal with
LLLLLR or RRRRRL.

These rotations cannot be performed if a node's parent is a member of
the verso spine.

Signed-off-by: Teodor-Dumitru Ene <teodord.ene@gmail.com>