Fix block implementation
First off, NetworkX's subgraph generation calls the __init__ of the
graph's class. The AdderTree for example. So all arguments to __init__
must be optional.

However, broken structure is not a concern, because the subgraph data is
just copied to a regular Graph afterwards. It just needs to initialize.

Further, the HDL function needs to check whether or not it is inside a
subgraph for the purpose of outputting corresponding HDL.

Signed-off-by: Teodor-Dumitru Ene <teodord.ene@gmail.com>