Improve code quality of tree unrank method
Previously, new nodes were added to the tree directly before a recursive
call. This led to redundant code that was hard to understand or improve.

Now, new nodes are added to the tree directly after a recursive call.

This also makes it easy to see that rank and unrank are opposites.

Signed-off-by: Teodor-Dumitru Ene <teodord.ene@gmail.com>