Remove ability to stereoscopically combine trees
I either lack the coding prowess or the theoretical understanding that
would be needed in order to implement the ability to stereoscopically
combine trees through methods that are both robust and efficient.

As such, the decision has been made to remove this option.

Instead of allowing users to attempt stereoscopic combination of
arbitrary tree structures, this library will enable users to apply such
optimizations in post-production.

Signed-off-by: Teodor-Dumitru Ene <teodord.ene@gmail.com>