Rehaul optimize_node functionality
Previously, optimize_node attempted to be operation-agnostic. It would
rely on there being multiple variants of each node, grouped by
footprint. It would then try to swap in the highest-priority nodes that
were valid.

This was a nightmare to debug, because it required methods to check for
whether long chains of nodes could be swapped into the tree to improve
the HDL.

Instead, the new method simply requires each operation (i.e: AdderTree)
to define what optimizations may be performed.

Closes #46

Signed-off-by: Teodor-Dumitru Ene <teodord.ene@gmail.com>